requirements.txt
tensorflow
rdkit-pypi
pandas==1.3.5
numpy
Pillow
matplotlib
pydot
scikit-learn==1.0.2
networkx
graphviz